@@ -20,31 +20,34 @@
   "homepage": "https://github.com/YongYuH/Ludo#readme",
   "dependencies": {
     "axios": "^0.13.1",
+    "blueimp-md5": "^2.6.0",
     "chart.js": "^2.3.0-rc.1",
+    "formsy-react": "^0.19.0",
     "imports-loader": "^0.6.5",
     "masonry-layout": "^4.1.0",
-    "material-ui": "^0.16.0",
+    "material-ui": "^0.16.4",
     "rc-progress": "^2.0.1",
     "rc-slider": "^4.0.1",
-    "react": "^15.2.0",
+    "react": "^15.3.2",
     "react-chartjs-2": "^1.2.5",
     "react-circular-progressbar": "^0.1.2",
-    "react-dom": "^15.2.0",
+    "react-dom": "^15.3.2",
     "react-dropzone": "^3.6.0",
-    "react-flipcard": "^0.2.1",
     "react-image-lightbox": "^3.1.0",
     "react-masonry-component": "^4.2.1",
     "react-notification-system": "^0.2.10",
     "react-router": "^2.5.2",
     "react-tag-input": "^4.3.3",
     "react-tagsinput": "^3.13.5",
-    "react-tap-event-plugin": "^1.0.0",
+    "react-tap-event-plugin": "^2.0.0",
     "react-textarea-autosize": "^4.0.5",
+    "react-ux-password-field": "^0.9.12",
     "react-widgets": "^3.4.4"
   },
   "devDependencies": {
     "babel-core": "^6.10.4",
     "babel-loader": "^6.2.4",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-react": "^6.11.1",
     "babel-preset-stage-0": "^6.16.0",
@@ -61,8 +64,8 @@
     "sass-loader": "^4.0.0",
     "style-loader": "^0.13.1",
     "url-loader": "^0.5.7",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1",
+    "webpack": "^1.13.2",
+    "webpack-dev-server": "^1.16.2",
     "webpack-merge": "^0.14.1",
     "webpack-validator": "^2.2.7"
   }
