@@ -9,17 +9,16 @@ export default class ActiveForPlayer extends React.Component {
     }
 
     componentWillMount() {
-        // console.log('ActiveForPlayer componentWillMount');   // debug
         this.props.handleIsOpeningActivePage(true);
     }
 
     componentWillUnmount() {
-        // console.log('ActiveForPlayer componentWillUnmount');   // debug
         this.props.handleIsOpeningActivePage(false);
     }
 
     render() {
         return (
+            /* components/_report-form.scss */
             <div className="form-and-report">
                 <ActivePlayerForm  {...this.props} />
                 <ActiveReports {...this.props} />
