@@ -38,14 +38,14 @@ export default class ActivePlayerForm extends React.Component {
             timeLineMarks: {},
             uploadImageIndex: 0
         };
+        this.handleCloseLightbox = this.handleCloseLightbox.bind(this);
         this.handleDayPickerClass = this.handleDayPickerClass.bind(this);
+        this.handleImageDrop = this.handleImageDrop.bind(this);
         this.handleImageEnlarge = this.handleImageEnlarge.bind(this);
         this.handleImageRemove = this.handleImageRemove.bind(this);
         this.handleReportTextChange = this.handleReportTextChange.bind(this);
         this.handleSubmit = this.handleSubmit.bind(this);
         this.handleTagsChange = this.handleTagsChange.bind(this);
-        this.onDrop = this.onDrop.bind(this);
-        this.closeLightbox = this.closeLightbox.bind(this);
         // this.moveNext = this.moveNext.bind(this);
         // this.movePrev = this.movePrev.bind(this);
     }
@@ -56,18 +56,15 @@ export default class ActivePlayerForm extends React.Component {
             this.setState({
                 isReportButtonClickable: true
             });
-            // console.log('componentWillReceiveProps getTimeLineMarks');   // debug
             this.getTimeLineMarks(nextProps);
         }
     }
 
     componentWillUnmount() {
-        // console.log('ActivePlayerForm componentWillUnmount');   // debug
         this.props.clearCurrentFormValue();
     }
 
     getTimeLineMarks(nextProps) {
-        // console.log('ActivePlayerForm getTimeLineMarks nextProps', nextProps);   // debug
         const { state } = this;
         const currentFormValue = nextProps.router_currentFormValue;
         const { checkpoint, duration } = currentFormValue;
@@ -92,13 +89,51 @@ export default class ActivePlayerForm extends React.Component {
         });
     }
 
+    handleCloseLightbox() {
+        this.setState({
+            isImageLightBoxOpen: false
+        });
+    }
+
     handleDayPickerClass(value) {
         const { checkpoint } = this.props.currentFormValue;
         const index = checkpoint.indexOf(value);
         if (index != -1) {
-            return ` ludo-detail-information-day-picker__button--checkpoint`;
+            return ' ludo-detail-information-day-picker__button--checkpoint';
         } else {
-            return ` ludo-detail-information-day-picker__button--duration`;
+            return ' ludo-detail-information-day-picker__button--duration';
+        }
+    }
+
+    handleImageDrop(files) {
+        if (files.length === 1) {
+            this.setState({
+                files,
+                isImageUploaded: true
+            });
+            const { ludoId }= this.props.params;
+            const imgPost = new FormData();
+            imgPost.append('file', files[0]);
+            axios.post('/apis/report-image', imgPost)
+            .then((response) => {
+                if (response.data.status == '200') {
+                    this.setState({
+                        imageLocation: response.data.location
+                    });
+                    // TODO: shouldComponentUpdate false when successfully upload an image
+                } else {
+                    console.error('image upload message from server: ', response.data.message);
+                    console.error('image upload error from server: ', response.data.err);
+                }
+            })
+            .catch((error) => {
+                console.error('image upload error', error);
+            });
+        } else if (files.length > 1) {
+            this.setState({
+                files: []
+            });
+            window.alert('一次只能上傳一張圖片');
         }
     }
 
@@ -110,7 +145,6 @@ export default class ActivePlayerForm extends React.Component {
 
     handleImageRemove(event) {
         event.preventDefault();
-        // const imageIndex = Number(event.currentTarget.value);  // For multiple picture upload
         this.setState({
             files: [],
             isImageUploaded: false
@@ -132,104 +166,68 @@ export default class ActivePlayerForm extends React.Component {
 
     handleSubmit(event) {
         event.preventDefault();
-        // TODO: Use notification confirming report
-        this.setState({
-            isReportButtonClickable: false
-        });
-        const isSureToReport = window.confirm(`Are you sure to report?`);
-        if (isSureToReport) {
-            const { ludo_id }= this.props.params;
-            const { currentUserId } = this.props;
-            const currentFormValue = this.props.router_currentFormValue;
-            let whoIsUser = '';
-            (currentFormValue.starter_id == currentUserId) ? whoIsUser = 'starter_check' : whoIsUser = 'player_check'
-            const ludoReportPost = {
-                'ludo_id': ludo_id,
-                'player': whoIsUser,
-                'content': this.state.reportText,
-                'image_location': this.state.imageLocation
-            };
-            // console.log('ludoTextReportPost', ludoReportPost);
-            console.log('ActivePlayerForm before report post');
-            axios.post('apis/report', ludoReportPost)
-            .then(response => {
-                if (response.data.status === '200') {
-                    this.props.handleShouldProfileUpdate(true);
-                    this.props.handleShouldReportUpdate(true);
-                    this.setState({
-                        isImageUploaded: false,
-                        reportText: ''
-                    });
-                    console.log('ActivePlayerForm Report after report post response.data: ', response.data);
-                } else {
+        // TODO: Use material-ui dialog confirming report
+        if (this.state.reportText || this.state.isImageUploaded) {
+            this.setState({
+                isReportButtonClickable: false
+            });
+            // const isSureToReport = window.confirm('Are you sure to report?');
+            const isSureToReport = window.confirm('確定要回報嗎？');
+            if (isSureToReport) {
+                const { ludo_id }= this.props.params;
+                const { currentUserId } = this.props;
+                const currentFormValue = this.props.router_currentFormValue;
+                let whoIsUser = '';
+                (currentFormValue.starter_id == currentUserId) ? whoIsUser = 'starter_check' : whoIsUser = 'player_check'
+                const ludoReportPost = {
+                    'ludo_id': ludo_id,
+                    'player': whoIsUser,
+                    'content': this.state.reportText,
+                    'image_location': this.state.imageLocation,
+                    'tags': this.state.reportTags
+                };
+                axios.post('apis/report', ludoReportPost)
+                .then((response) => {
+                    if (response.data.status === '200') {
+                        this.props.handleShouldProfileUpdate(true);
+                        this.props.handleShouldReportUpdate(true);
+                        this.setState({
+                            files: [],
+                            isImageUploaded: false,
+                            reportTags: [],
+                            reportText: ''
+                        });
+                    } else {
+                        window.alert('回報時發生錯誤，請重試一次；若問題還是發生，請聯絡開發人員');
+                        console.error('ActivePlayerForm Report else message from server: ', response.data.message);
+                        if (response.data.err) {
+                            console.error('ActivePlayerForm Report else error from server: ', response.data.err);
+                        }
+                        this.setState({
+                            isReportButtonClickable: true
+                        });
+                    }
+                })
+                .catch((error) => {
                     window.alert('回報時發生錯誤，請重試一次；若問題還是發生，請聯絡開發人員');
-                    console.log('ActivePlayerForm Report else message from server: ', response.data.message);
-                    console.log('ActivePlayerForm Report else error from server: ', response.data.err);
+                    console.error('ActivePlayerForm Report error', error);
                     this.setState({
                         isReportButtonClickable: true
                     });
-                }
-            })
-            .catch(error => {
-                window.alert('回報時發生錯誤，請重試一次；若問題還是發生，請聯絡開發人員');
-                console.log('ActivePlayerForm Report error', error);
+                });
+            } else {
                 this.setState({
                     isReportButtonClickable: true
                 });
-            });
+            }
         } else {
-            this.setState({
-                isReportButtonClickable: true
-            });
+            window.alert('你尚未輸入回報內容或上傳圖片');
         }
     }
 
     handleTagsChange(reportTags) {
         this.setState({
             reportTags
-        });
-    }
-
-    onDrop(files) {
-        console.log('onDrop files', files);
-        if (files.length == 1) {
-            this.setState({
-                files,
-                isImageUploaded: true
-            });
-            const { ludoId }= this.props.params;
-            const imgPost = new FormData();
-            imgPost.append('file', files[0]);
-            // console.log('imgPost file', imgPost.get('file'));
-            console.log('before post image');
-            axios.post('/apis/report-image', imgPost)
-            .then(response => {
-                if (response.data.status == '200') {
-                    // window.alert(`image upload success!`);
-                    // console.log('image upload response.data: ', response.data);
-                    this.setState({
-                        imageLocation: response.data.location
-                    });
-                    // TODO: shouldComponentUpdate false when successfully upload an image
-                } else {
-                    console.log('image upload message from server: ', response.data.message);
-                    console.log('image upload error from server: ', response.data.err);
-                }
-            })
-            .catch(error => {
-                console.log('image upload error', error);
-            });
-        } else if (files.length > 1) {
-            this.setState({
-                files: []
-            });
-            window.alert('一次只能上傳一張圖片');
-        }
-    }
-
-    closeLightbox() {
-        this.setState({
-            isImageLightBoxOpen: false
         });
     }
 
@@ -252,20 +250,27 @@ export default class ActivePlayerForm extends React.Component {
     // }
 
     render() {
-        // const { currentFormValue } = this.props;
         const currentFormValue = this.props.router_currentFormValue;
         const { ludoDetailInformation, category, files, 
             isImageLightBoxOpen, isImageUploaded, isReportButtonClickable, isReportTextBlank, 
-            maxDuration, maxMarbles, timeLineMarks, uploadImageIndex 
+            maxDuration, maxMarbles, reportText, timeLineMarks, uploadImageIndex 
         } = this.state;
-        const { category_id, checkpoint, duration, introduction, reportText, marbles, tags, title } = currentFormValue;
+        const { category_id, checkpoint, duration, introduction, marbles, tags, title } = currentFormValue;
         const dayPickerButtons = [];
         return (
+            /* components/_report-form.scss */
             <div className="form--report">
-                <form onSubmit={this.handleSubmit} className="ludo-detail-information-container report-information-container">
+                {/* components/_ludo-detail-information.scss */}
+                <form
+                    className="ludo-detail-information-container report-information-container"
+                    onSubmit={this.handleSubmit}
+                >
                     <div className="ludo-detail-information-top-container">
                         <div className="category-icon-container">
-                            <img className="category-icon" src={iconArray[category_id]} />
+                            <img
+                                className="category-icon"
+                                src={iconArray[category_id]}
+                            />
                         </div>
                         <div className="top-right-container">
                             <div className="text-field-container">
@@ -280,16 +285,22 @@ export default class ActivePlayerForm extends React.Component {
                                     {title}
                                 </span>
                             </div>
+                            {/* components/_marbles.scss */}
                             <div className="label-and-slider">
                                 <div className="text-label">
                                     彈珠數:<span className="text-label--marble-number">{marbles}</span>
                                 </div>
                                 <div className="ludo-detail-information-slider--marbles">
-                                    <RcSlider max={maxMarbles} value={marbles} disabled={true} />
+                                    <RcSlider
+                                        disabled
+                                        max={maxMarbles}
+                                        value={marbles}
+                                    />
                                 </div>
                             </div>
                         </div>
                     </div>
+                    {/* components/_report-form.scss */}
                     <div className="report-form-bottom-container report-form-bottom-container--player">
                         <div className="introduction-and-duration">
                             <div className="label-and-introduction--player">
@@ -298,18 +309,19 @@ export default class ActivePlayerForm extends React.Component {
                                     <div className="introduction">
                                         {introduction}
                                     </div>
+                                    {/* components/_tags.scss */}
                                     <div className="text-field--hashtag">
                                         <div className="react-tagsinput">
                                             <span className="react-tagsinput-span">
                                                 {
                                                     tags.length ?
-                                                    tags.map((tagString, index) => {
-                                                        return (
-                                                            <span className="react-tagsinput-tag" key={`tag-${index}`}>
-                                                                {tagString}
-                                                            </span>
-                                                        );
-                                                    })
+                                                        tags.map((tagString, index) => {
+                                                            return (
+                                                                <span className="react-tagsinput-tag" key={`tag-${index}`}>
+                                                                    {tagString}
+                                                                </span>
+                                                            );
+                                                        })
                                                     : null
                                                 }
                                             </span>
@@ -321,9 +333,14 @@ export default class ActivePlayerForm extends React.Component {
                                 <div className="text-label">持續期間:</div>
                                 <div className="report-time-line-container">
                                     <div className="report-time-line">
-                                    <RcSlider className="time-line" disabled={true} vertical={true} dots included={false}
-                                        max={duration} min={1} value={checkpoint} range={checkpoint.length}
+                                    <RcSlider
+                                        className="time-line"
+                                        disabled vertical dots included={false}
                                         marks={timeLineMarks}
+                                        max={duration}
+                                        min={1}
+                                        range={checkpoint.length}
+                                        value={checkpoint}
                                     />
                                     </div>
                                 </div>
@@ -332,11 +349,11 @@ export default class ActivePlayerForm extends React.Component {
                             <div className="upload-container">
                                 <div className="upload-picture-button-container">
                                     <DropZone 
+                                        accept={"image/*"}
                                         className="upload-picture-button"
                                         maxSize={2000000}
-                                        onDrop={this.onDrop}
-                                        onClick={this.onDrop}
-                                        accept={"image/png", "image/pjepg", "image/jpeg"}
+                                        onClick={this.handleImageDrop}
+                                        onDrop={this.handleImageDrop}
                                     >
                                         <img className="upload-picture-button__icon" src={uploadIcon}/>
                                     </DropZone>
@@ -344,21 +361,21 @@ export default class ActivePlayerForm extends React.Component {
                                 <div className="upload-content-and-tags">
                                     <textarea 
                                         className="upload-text-container"
+                                        maxLength="140"
+                                        onChange={this.handleReportTextChange}
                                         // placeholder="Report here"
                                         placeholder="輸入要回報的內容，140字為限"
                                         rows="6"
-                                        maxLength="140"
                                         value={isReportTextBlank ? '' : this.state.reportText}
-                                        onChange={this.handleReportTextChange}
                                     />
                                     {
                                         isImageUploaded ?
                                             <div className="upload-preview">
-                                                {/* <span className="upload-preview__text">正要上傳的圖片: </span> */}
                                                 <div className="upload-preview__image-container">
-                                                    <img className="upload-preview__image" 
-                                                        src={files[uploadImageIndex].preview}
+                                                    <img
+                                                        className="upload-preview__image"
                                                         onClick={this.handleImageEnlarge}
+                                                        src={files[uploadImageIndex].preview}
                                                     />
                                                     <div className="upload-preview-instruction-container">
                                                         <button 
@@ -375,28 +392,30 @@ export default class ActivePlayerForm extends React.Component {
                                     }
                                     <div className="upload-report-tags-container">
                                         <TagsInput
-                                            value={this.state.reportTags} 
-                                            onChange={this.handleTagsChange}
                                             inputProps={{maxLength: 30, placeholder:"標籤"}}
+                                            onChange={this.handleTagsChange}
+                                            value={this.state.reportTags} 
                                         />
                                     </div>
                                 </div>
                             </div>
                     </div>
-                        <button className="report-submit-button report-submit-button--report" type="submit" 
-                            disabled={isReportTextBlank && !isImageUploaded && !isReportButtonClickable}
-                        >
-                            回報
-                        </button>
+                    {/* components/_submit-button.scss */}
+                    <button
+                        className="report-submit-button report-submit-button--report"
+                        disabled={isReportTextBlank && !isImageUploaded || !isReportButtonClickable}
+                        type="submit"
+                    >
+                        回報
+                    </button>
                 </form>
                 {
                     isImageLightBoxOpen ? 
                         <Lightbox 
-                            className="lighbox-target"
                             mainSrc={files[uploadImageIndex].preview}
                             // nextSrc={files.length == 1 ? null : files[(uploadImageIndex + 1) % files.length].preview}
                             // prevSrc={files.length == 1 ? null : files[(uploadImageIndex + files.length - 1) % files.length].preview}
-                            onCloseRequest={this.closeLightbox}
+                            onCloseRequest={this.handleCloseLightbox}
                             // onMovePrevRequest={this.movePrev}
                             // onMoveNextRequest={this.moveNext}
                         />
@@ -406,26 +425,3 @@ export default class ActivePlayerForm extends React.Component {
         );
     }
 };
-                    // {
-                    //     files.length > 0 ? 
-                    //         <div className="upload-instruction">
-                    //             Ready to upload 
-                    //             <span className="number">{files.length}</span>
-                    //             image
-                    //             {files.length > 1 ? <span>s</span> : null}
-                    //         </div>
-                    //     : null
-                    // }
-
-                            // <div className="upload-picture-button-container">
-                            //     <span className="upload-hint-text">140字為限(#tag不在此限)</span>
-                            //     <DropZone 
-                            //         className="upload-picture-button"
-                            //         maxSize={2000000}
-                            //         onDrop={this.onDrop}
-                            //         onClick={this.onDrop}
-                            //         accept={"image/png", "image/pjepg", "image/jpeg"}
-                            //     >
-                            //         <img className="upload-picture-button__icon" src={uploadIcon}/>
-                            //     </DropZone>
-                            // </div>
