@@ -7,65 +7,80 @@ import ActiveForPlayer from '../active/active-for-player/ActiveForPlayer';
 import App from './App';
 import Create from '../create/Create';
 import Friend from '../friend/Friend';
-import Login from './Login';
+import Invite from '../create/Invite';
+import LogIn from './LogIn.js';
 import LudoEdit from '../ludo-edit/LudoEdit';
 import OpenedForStarter from '../opened/opened-for-starter/OpenedForStarter';
 import OpenedForBystander from '../opened/opened-for-bystander/OpenedForBystander';
 import Playground from '../playground/Playground';
 import Profile from '../profile/Profile';
+import SignUp from './SignUp';
+import Template from '../create/Template';
 
-const ludoPageArray = [OpenedForBystander, OpenedForStarter, OpenedForBystander, ActiveForPlayer, ActiveForPlayer, ActiveForBystander];
-const ludoPageArrayForEdit = [OpenedForBystander, LudoEdit, OpenedForBystander, ActiveForPlayer, ActiveForPlayer, ActiveForBystander];
+/*
+    auth        statement：
+    0           not login
+    1           stage 1 creator
+    2           stage 1 other players
+    3           stage 2 playerA (creator)
+    4           stage 2 playerB
+    5           stage 2 other players
+    6           stage 3 all
+    7           stage 0 creator
+    8           stage 0 other players
+    9           stage 0 not login (same as 0, may be modified in the future)
+*/
 
-// import '../../stylesheets/main.scss';
+const ludoPageArray = [OpenedForBystander, OpenedForStarter, OpenedForBystander, ActiveForPlayer, ActiveForPlayer, ActiveForBystander, ActiveForBystander, Template, Template, Template];
+const ludoPageArrayForEdit = [OpenedForBystander, LudoEdit, OpenedForBystander, ActiveForPlayer, ActiveForPlayer, ActiveForBystander, ActiveForBystander, Template, Template, Template];
 
 const isLoggedIn = (nextState, replace, callback) => {
-    // TODO: Look up the detail usage of replace function
+    /* TODO: Look up the detail usage of replace function */
     axios.get('/apis/user')
-    .then(response => {
+    .then((response) => {
         if(response.data.status != '200') {
-            replace(`/login`);
+            replace('/login');
         }
         callback();
     })
-    .catch(error => {
-        console.log('AppRouter isloggedin error');
+    .catch((error) => {
+        console.error('AppRouter isloggedin error');
         callback(error);
     })
 };
 
 let router_ludoPageIndex = null;
 let router_currentFormValue = {};
+let router_currentLudoId = '';
 
 const ludoRedirect = (nextState, replace, callback) => {
     const { ludo_id }= nextState.params;
-    // console.log('ludoRedirect before get single ludo -- ludo_id', ludo_id);  // debug
     axios.get(`/apis/ludo/${ludo_id}`)
-    .then(response => {
+    .then((response) => {
         if(response.data.status === '200') {
             router_ludoPageIndex = response.data.auth;
-            // console.log('router_ludoPageIndex', router_ludoPageIndex);   // debug
             router_currentFormValue = response.data.ludo;
-            // console.log('router_currentFormValue', router_currentFormValue);   // debug
+            router_currentLudoId = response.data.ludo.ludo_id;
             callback();
         } else {
-            // console.log('AppRouter ludoRedirect getCurrentLudoData else response from server: ', response);
             window.alert('取得Ludo時發生錯誤，請重試一次；若問題還是發生，請聯絡開發團隊');
-            console.log('AppRouter ludoRedirect getCurrentLudoData else message from server: ', response.data.message);
-            console.log('AppRouter ludoRedirect getCurrentLudoData else error from server: ', response.data.err);
+            console.error('AppRouter ludoRedirect getCurrentLudoData else response from server: ', response);
+            console.error('AppRouter ludoRedirect getCurrentLudoData else message from server: ', response.data.message);
+            if (response.data.err) {
+                console.error('AppRouter ludoRedirect getCurrentLudoData else error from server: ', response.data.err);
+            }
         }
     })
-    .catch(error => {
+    .catch((error) => {
         window.alert('取得Ludo時發生錯誤，請重試一次；若問題還是發生，請聯絡開發團隊');
-        console.log('AppRouter ludoRedirect getCurrentLudoData error', error);
+        console.error('AppRouter ludoRedirect getCurrentLudoData error', error);
     });
 };
 
 const ludoEditRedirect = (nextState, replace, callback) => {
     const { ludo_id }= nextState.params;
-    // console.log('ludoRedirect before get single ludo -- ludo_id', ludo_id);  // debug
     axios.get(`/apis/ludo/${ludo_id}`)
-    .then(response => {
+    .then((response) => {
         if(response.data.status === '200') {
             if(response.data.auth != 1) {
                 /* TODO: Figure out how to use same url redirect to other component */
@@ -73,21 +88,22 @@ const ludoEditRedirect = (nextState, replace, callback) => {
                 browserHistory.push(`/ludo/${ludo_id}`);
             } else {
                 router_ludoPageIndex = response.data.auth;
-                // console.log('router_ludoPageIndex', router_ludoPageIndex);   // debug
                 router_currentFormValue = response.data.ludo;
-                // console.log('router_currentFormValue', router_currentFormValue);   // debug
+                router_currentLudoId = response.data.ludo.ludo_id;
                 callback();
             }
         } else {
-            // console.log('AppRouter ludoRedirect getCurrentLudoData else response from server: ', response);
             window.alert('取得Ludo時發生錯誤，請重試一次；若問題還是發生，請聯絡開發團隊');
-            console.log('AppRouter ludoRedirect getCurrentLudoData else message from server: ', response.data.message);
-            console.log('AppRouter ludoRedirect getCurrentLudoData else error from server: ', response.data.err);
+            console.error('AppRouter ludoRedirect getCurrentLudoData else response from server: ', response);
+            console.error('AppRouter ludoRedirect getCurrentLudoData else message from server: ', response.data.message);
+            if (response.data.err) {
+                console.error('AppRouter ludoRedirect getCurrentLudoData else error from server: ', response.data.err);
+            }
         }
     })
-    .catch(error => {
+    .catch((error) => {
         window.alert('取得Ludo時發生錯誤，請重試一次；若問題還是發生，請聯絡開發團隊');
-        console.log('AppRouter ludoRedirect getCurrentLudoData error', error);
+        console.error('AppRouter ludoRedirect getCurrentLudoData error', error);
     });
 };
 
@@ -100,31 +116,43 @@ export default class AppRouter extends React.Component {
                 <Route path="/" component={App}>
                     <IndexRedirect to="playground"></IndexRedirect>
                     <Route path="create" component={Create} onEnter={isLoggedIn}></Route>
-                    <Route path="friend" component={Friend}></Route>
-                    <Route path="login" component={Login}></Route>
-                    <Route path="ludo/:ludo_id" getComponent={(nextState, cb) => {
-                        const Component = ludoPageArray[router_ludoPageIndex];
-                        cb(null, props => <Component {...props} router_currentFormValue={router_currentFormValue} />);
-                    }} onEnter={ludoRedirect}></Route>
-                    <Route path="ludo-edit/:ludo_id" getComponent={(nextState, cb) => {
-                        const Component = ludoPageArrayForEdit[router_ludoPageIndex];
-                        cb(null, props => <Component {...props} router_currentFormValue={router_currentFormValue} />);
-                    }} onEnter={isLoggedIn, ludoEditRedirect} OnLeave={ludoRedirect}></Route>
+                    <Route path="friend" component={Friend} onEnter={isLoggedIn}></Route>
+                    <Route path="invite/:friend_id" component={Invite} onEnter={isLoggedIn}></Route>
+                    <Route path="login" component={LogIn}></Route>
+                    <Route
+                        path="ludo/:ludo_id"
+                        getComponent={(nextState, cb) => {
+                            const Component = ludoPageArray[router_ludoPageIndex];
+                            cb(null, props => 
+                                <Component 
+                                    {...props} 
+                                    router_currentFormValue={router_currentFormValue}
+                                    router_currentLudoId={router_currentLudoId}
+                                />);
+                        }} 
+                        onEnter={ludoRedirect}
+                    >
+                    </Route>
+                    <Route
+                        path="ludo-edit/:ludo_id"
+                        getComponent={(nextState, cb) => {
+                            const Component = ludoPageArrayForEdit[router_ludoPageIndex];
+                            cb(null, props => 
+                                <Component 
+                                    {...props}
+                                    router_currentFormValue={router_currentFormValue}
+                                    router_currentLudoId={router_currentLudoId}
+                                />);
+                        }} 
+                        onEnter={isLoggedIn, ludoEditRedirect}
+                        OnLeave={ludoRedirect}
+                    >
+                    </Route>
                     <Route path="playground" component={Playground}></Route>
-                    <Route path="profile/:userId" component={Profile} onEnter={isLoggedIn}></Route>
+                    <Route path="profile(/:userId)" component={Profile} onEnter={isLoggedIn}></Route>
+                    <Route path="signup" component={SignUp}></Route>
                 </Route>
             </Router>
         );
     }
-};
-
-                    // <Route path="active-for-bystander/:ludoId" component={ActiveForBystander}></Route>
-                    // <Route path="active-for-player/:ludoId" component={ActiveForPlayer}></Route>
-                    // <Route path="create" component={Create} onEnter={isLoggedIn}></Route>
-                    // <Route path="friend" component={Friend}></Route>
-                    // <Route path="login" component={Login}></Route>
-                    // <Route path="ludo-edit/:ludo_id" component={LudoEdit} onEnter={isLoggedIn}></Route>
-                    // <Route path="opened-for-starter/:ludoId" component={OpenedForStarter} onEnter={isLoggedIn}></Route>
-                    // <Route path="opened-for-bystander/:ludoId" component={OpenedForBystander}></Route>
-                    // <Route path="playground" component={Playground}></Route>
-                    // <Route path="profile/:userId" component={Profile}></Route>
+}
