@@ -1,118 +1,130 @@
 import React from "react";
 import axios from '../axios-config';
+import { browserHistory, Link } from 'react-router';
+import Formsy from 'formsy-react';
+import md5 from 'blueimp-md5';
+
+import FormsyHOCInput from './FormsyHOCInput.js';
+import PasswordField from './PasswordField.js';
 
 import logoImgPath from '../../images/Ludo_logo.png';
-import googleIcon from '../../images/login/google-icon.png';
-import twitterIcon from '../../images/login/twitter-icon.png';
+// import googleIcon from '../../images/login/google-icon.png';
+// import twitterIcon from '../../images/login/twitter-icon.png';
 import facebookIcon from '../../images/login/facebook-icon.png';
 
-export default class Login extends React.Component {
-    constructor() {
-        super();
+export default class LogIn extends React.Component {
+    constructor(props) {
+        super(props);
         this.state = {
-            loginInformation: {
-                email: '',
-                password: ''
-            },
-            isEmailBlank: true,
-            isPasswordBlank: true
+            canSubmit: false,
+            errorMessageFromServer: ''
         };
-        this.handleEmailField = this.handleEmailField.bind(this);
-        this.handlePasswordField = this.handlePasswordField.bind(this);
-        this.handleLogin = this.handleLogin.bind(this);
+        this.disableButton = this.disableButton.bind(this);
+        this.enableButton = this.enableButton.bind(this);
+        this.handleLogIn = this.handleLogIn.bind(this);
     }
 
-    handleEmailField(event) {
-        if(event.target.value) {
-            this.setState(
-                Object.assign(this.state, {
-                    isEmailBlank: false
-                })
-            );
-        } else {
-            this.setState( 
-                Object.assign(this.state, {
-                    isEmailBlank: true
-                })
-            );
-        }
+    disableButton() {
+        this.setState({
+            canSubmit: false
+        });
     }
 
-    handlePasswordField(event) {
-        if(event.target.value) {
-            this.setState(
-                Object.assign(this.state, {
-                    isPasswordBlank: false
-                })
-            );
-        } else {
-            this.setState(
-                Object.assign(this.state, {
-                    isPasswordBlank: true
-                })
-            );
-        }
+    enableButton() {
+        this.setState({
+            canSubmit: true
+        });
     }
 
-    handleLogin() {
-        const { isEmailBlank, isPasswordBlank } = this.state;
-        if (!isEmailBlank && !isPasswordBlank) {
-            const { loginInformation } = this.state;
-            axios.post('/login', loginInformation)
-            .then(function (response) {
-                window.alert(`successfully login`);
-            })
-            .catch(function (error) {
-                console.log('error', error);
-            });
-        } else {
-            window.alert(`Lack of email or password`);
-        }
+    handleSignUp(event) {
+        browserHistory.push('/playground');
+    }
+
+    handleLogIn(logInData) {
+        logInData.password = md5(logInData.password);
+        axios.post('/login', logInData)
+        .then((response) => {
+            if (response.data.status === '200') {
+                this.props.handleShouldUserBasicDataUpdate(true);
+                browserHistory.push('/playground');
+            // } else if (response.data.status === '') {
+                 /* wrong password */
+            //     console.log('wrong password');
+            // } else if (response.data.status === '') {
+                 /* email not exist */
+            //     console.log('email not exist');
+            } else {
+                this.setState({
+                    errorMessageFromServer: response.data.message[0]
+                });
+                console.error('Login handleLogIn response', response);
+            }
+        })
+        .catch((error) => {
+            console.error('Login handleLogIn error', error);
+        });
     }
 
     render() {
-        return ( 
-            <div className="grid">
-                <div className="login-container">
-                    <div className="login-top-container">
-                        <div className="logo-container">
-                            <img src={logoImgPath} className="logo-icon" />
-                        </div>
-                        <div className="email">
-                            <input type="text" placeholder="email" onChange={this.handleEmailField}/>
-                        </div>
-                        <div className="password">
-                            <input type="text" placeholder="password" onChange={this.handlePasswordField}/>
-                        </div>
-                        <button className="login" onClick={this.handleLogin}>
-                            Log in
-                        </button>
-                        <button className="signup">
-                            Sign up
-                        </button>
+        return (
+            /* components/_login.scss */
+            <div className="login-container">
+                <Formsy.Form
+                    className="login-top-container"
+                    onValid={this.enableButton}
+                    onValidSubmit={this.handleLogIn}
+                >
+                    <div className="logo-container">
+                        <img src={logoImgPath} className="logo-icon" />
                     </div>
-                    <div className="login-bottom-container">
-                        <div className="other-login-instruction">
-                            <p>or</p>
-                            <br />
-                            <p>Log in via your favorite social media</p>
-                        </div>
-                        <div className="other-login-icon-container">
-                            <div className="twitter">
-                                <img src={twitterIcon} className="other-login-icon" />
-                            </div>
-                            <div className="google">
-                                <img src={googleIcon} className="other-login-icon" />
-                            </div>
-                            <div className="facebook">
-                                <a href="http://api.ludonow.com/auth/facebook">
-                                    <img src={facebookIcon} className="other-login-icon" />
-                                </a>
-                            </div>
-                        </div>
+                    <FormsyHOCInput
+                        className="email"
+                        name="email"
+                        placeholder="輸入Email"
+                        required
+                        validations="isEmail"
+                    />
+                    <PasswordField
+                        className="password"
+                        name="password"
+                        placeholder="輸入密碼"
+                        required
+                        validations="isPassword"
+                    />
+                    <div className="server-error-message">
+                        {this.state.errorMessageFromServer}
                     </div>
+                    <div className="buttons">
+                        <button
+                            className="login"
+                            type="submit"
+                        >
+                            登入
+                        </button>
+                        <Link
+                            className="signup"
+                            to="signup"
+                        >
+                            註冊
+                        </Link>
+                    </div>
+                    <a
+                        className="forget-password"
+                        href="http://api.ludonow.com/users/password/new"
+                    >
+                        忘記密碼
+                    </a>
+                </Formsy.Form>
+                <div className="login-bottom-container">
+                    <a href="http://api.ludonow.com/auth/facebook">
+                        <img
+                            className="fb-login-icon"
+                            src={facebookIcon}
+                        />
+                    </a>
+                    <div className="fb-login-text">FB登入</div>
                 </div>
             </div>
         );
     }
-};
+}
