@@ -11,12 +11,12 @@ import unmentionablesIcon from '../../images/category_icon/unmentionables.png';
 import othersIcon from '../../images/category_icon/others.svg';
 const iconArray = [othersIcon, lifestyleIcon, readIcon, exerciseIcon, studyIcon, newSkillIcon, unmentionablesIcon, othersIcon];
 import marbleIcon from '../../images/marble.png';
-
+const category= ['其它', '生活作息', '閱讀', '運動', '教科書', '新技能', '不可被提起的', '其它']
 const categoryClassArray = [' others', ' lifestyle', ' read', ' exercise', ' study', ' new_skill', ' unmentionables', ' others'];
 
 const masonryOptions = {
     itemSelector: ".form-item",
-    columnWidth: 220,
+    columnWidth: 226,
     fitWidth: true
 };
 
@@ -44,27 +44,23 @@ export default class LudoList extends React.Component {
     }
 
     handleCardStage(stage) {
-        if (stage == 1) {
-            return `card-bottom__stage--opened`;
-        } else {
-            return `card-bottom__stage--closed`;
-        }
-    }
-
-    handleCardBackClass(category_id) {
-        return categoryClassArray[category_id];
-    }
-
-    handleCardBottomGoClass(category_id) {
-        return categoryClassArray[category_id];
-    }
-
-    handleCardFrontTopClass(category_id) {
-        return categoryClassArray[category_id];
-    }
-
-    handleCategoryIcon(category_id) {
-        return iconArray[category_id];
+      switch (stage) {
+          case 0: 
+              return 'card-top__stage--0';
+              break;
+          case 1:
+              return 'card-top__stage--1';
+              break;
+          case 2:
+              return 'card-top__stage--2';
+              break;
+          case 3:
+              return 'card-top__stage--3';
+              break;
+          default:
+              return 'card-top__stage--0';
+              break;
+      }
     }
 
     handleCardLink(event) {
@@ -110,73 +106,101 @@ export default class LudoList extends React.Component {
         const { currentFormValue, shouldProfileUpdate, ludoList } = this.props;
         return (
             <Masonry options={masonryOptions}>
+
                 {
-                    ludoList.map( (singleLudoObject, index) => {
+                    ludoList.map((singleLudoObject, index) => {
                         const isThisCardFlipped = (this.state.flippedKey.indexOf(index) != -1);
                         const buttonClickHandler = isThisCardFlipped ? this.showFront : this.showBack;
                         return (
-                            <div className={`form-item`} key={`card-${index}`}>
-                                <div 
-                                    className={`form-card card-front ${isThisCardFlipped ? "" : "card-flip"}`}
+                            /* components/_form.scss */
+                             <div
+                                className="form-item"
+                                key={`card-${index}`}
+                            >
+                                {/* form-card is defined in components/_form.scss, others are defined in components/_card.scss */}
+                                <div
+                                    className={`card card--playground card-front ${isThisCardFlipped ? 'card-flip' : ''}`}
                                     id={index}
                                     onClick={buttonClickHandler}
                                 >
-                                    <div className={`card-top ${this.handleCardFrontTopClass(singleLudoObject.category_id)}`}>
-                                        <div className="title">{singleLudoObject.title}</div>
-                                        <div className="duration">{singleLudoObject.duration} days</div>
-                                        <div className="card-marble">
-                                            <img src={marbleIcon} className="card-marble__icon" />
-                                            <span className="card-marble__number">{singleLudoObject.marbles}</span>
-                                        </div>
-                                    </div>
-                                    <div className="card-bottom">
-                                        <img className="card-bottom__category-icon" src={this.handleCategoryIcon(singleLudoObject.category_id)} />
-                                        <div className={`card-bottom__stage ${this.handleCardStage(singleLudoObject.stage)}`} />
-                                    </div>
-                                </div>
-                                <div 
-                                    className={`form-card card-back card-back--form-list${isThisCardFlipped ? " card-flip" : ""} ${this.handleCardBackClass(singleLudoObject.category_id)}`}
-                                    id={index}
-                                    onClick={buttonClickHandler}
+                                <div className={`card-back ${categoryClassArray[singleLudoObject.category_id]}`}
+                                id={index}
+                                onClick={buttonClickHandler}
                                 >
-                                    <div className={this.handleCardBackClass(singleLudoObject.category_id)}>
-                                        <div className="card-introduction">
-                                            {
-                                                String(singleLudoObject.introduction).length > 20 ?
-                                                String(singleLudoObject.introduction).substring(0, 20) + ' ...'
-                                                : String(singleLudoObject.introduction)
-                                            }
-                                        </div>
-                                        <div className="card-hashtags">
-                                            {
-                                                // TODO: Use presentational component and proptypes to receive ludolist data
-                                                Array.isArray(singleLudoObject.tags) && singleLudoObject.tags ?
-                                                    singleLudoObject.tags.map( (tagString, tagIndex) => {
-                                                        return (
-                                                            <span className="react-tagsinput-tag--card" key={`tag-${tagIndex}`}>
-                                                                {tagString}
-                                                            </span>
-                                                        );
-                                                    })
-                                                : null
-                                            }
-                                        </div>
+                                {/*three information: star(for user to highlight this card), category, and introduction*/}
+                                  <div className="card-information">
+
+                                    <div className="card-star">
                                     </div>
-                                    <div className={`card-bottom ${this.handleCardBottomGoClass(singleLudoObject.category_id)}`}>
-                                        <div className={`card-bottom__triangle ${this.handleCardBottomGoClass(singleLudoObject.category_id)}`}>
-                                            <div className={`card-bottom__text ${this.handleCardBottomGoClass(singleLudoObject.category_id)}`}
-                                                onClick={this.handleCardLink} id={`go-${index}`}
-                                            >
-                                                go
-                                            </div>
-                                        </div>
+                                    <div className ="card-category">
+                                    {category[singleLudoObject.category_id]}
                                     </div>
+                                    <div className="card-introduction">{
+                                      String(singleLudoObject.introduction).length > 20 ?
+                                      String(singleLudoObject.introduction).substring(0, 20) + ' ...'
+                                      : String(singleLudoObject.introduction)
+                                      /* String(singleLudoObject.introduction)*/
+                                      }
+                                    </div>
+                                    <div className="card-hashtags">
+                                      {
+                                      // TODO: Use presentational component and proptypes to receive ludolist data
+                                      Array.isArray(singleLudoObject.tags) && singleLudoObject.tags ?
+                                          singleLudoObject.tags.map((tagString, tagIndex) => {
+                                              return (
+                                                  /* components/_tags.scss */
+                                                  <span
+                                                      className="react-tagsinput-tag--card"
+                                                      key={`tag-${tagIndex}`}
+                                                  >
+                                                      {tagString}
+                                                  </span>
+                                              );
+                                          })
+                                      : null
+                                    }
+                                    </div>
+                                  </div>
+                                    {/*the circle button for GO */}
+                                  <div className="card-button_circle">
+                                    <div className="card-button_text"
+                                      id={`go-${index}`}
+                                      onClick={this.handleCardLink}
+                                    >
+                                    Go
+                                    </div>
+                                  </div>
                                 </div>
+
+
+                                <div className={`card-top__stage ${this.handleCardStage(singleLudoObject.stage)}`} ></div>
+                                <div className="card-front-info">
+                                  <div className="card-star"></div>
+                                    <img  className="category-icon"
+                                        src={iconArray[singleLudoObject.category_id]}
+                                    />
+
+                                  <div className="title">{singleLudoObject.title}</div>
+                                  <div className="duration">{singleLudoObject.duration} days</div>
+                                  <div className="card-marble">
+                                    <img
+                                        className="card-marble__icon"
+                                        src={marbleIcon}
+                                        />
+                                    <span className="card-marble__number">{singleLudoObject.marbles}</span>
+                                  </div>
+
+                                </div>
+                                <div className="card-views">{singleLudoObject.views}</div>
+
+                              </div>
                             </div>
-                        )
+
+
+                        );
                     })
                 }
             </Masonry>
         );
     }
-};
+}
