@@ -3,10 +3,23 @@ import React from 'react';
 import FriendList from './FriendList';
 
 export default class Friend extends React.Component {
+    constructor(props) {
+        super(props);
+    }
+
+    componentWillMount() {
+        this.props.handleIsOpeningProfilePage(true);
+    }
+
+    componentWillUnmount() {
+        this.props.handleIsOpeningProfilePage(false);
+    }
+
     render() {
         return (
+            /* layout/main-container */
             <div className="main-container">
-                <FriendList />
+                <FriendList friendsArrayData={this.props.userProfileData.friends} />
             </div>
         );
     }
